+++ crates/ironcache-server/src/lib.rs
@@ -20,6 +20,7 @@ pub mod cmd_bitmap;
 pub mod cmd_block;
 pub mod cmd_cluster;
 pub mod cmd_config;
+pub mod cmd_dump;
 pub mod cmd_expire;
 pub mod cmd_hash;
 pub mod cmd_hll;